Don't pass all modified columns as event params
If one of the modified columns on the tracked object matches a unique
column on the tracking table it will do the Wrong Thing.  The easiest to
describe is if both tables have the same name for the PK, perhaps "id".
If you insert and then update a record, the tracking table will now have
a MAX(id) of 2 so if you try to insert another tracked item it will
insert id 2 and try to insert an event with the same id and violate the
unique constraint as well as not doing what you'd want.