Fix the delete wrapper with foreign key checks
If we enable foreign key checks, then we can't insert a 'delete' event
for a record that no longer exists.

Unfortunately, this removes the foreign key constraint from our event
tables, but I'm not sure that's wrong, since we really just want to be
able to group all these events together, and search for events from the
actual instance.